requirements.txt
openai
Pillow
python-dotenv
python-telegram-bot